@@ -1,48 +1,45 @@
 {
-    "name": "mist-webhook-monitor",
-    "version": "0.0.0",
-    "scripts": {
-        "ng": "ng",
-        "start": "ng serve --proxy-config ./proxy.conf.json --host 0.0.0.0",
-        "build": "ng build",
-        "watch": "ng build --watch --configuration development",
-        "test": "ng test"
-    },
-    "private": true,
-    "dependencies": {
-        "@angular/animations": "~13.4.0",
-        "@angular/common": "~13.4.0",
-        "@angular/compiler": "~13.4.0",
-        "@angular/core": "~13.4.0",
-        "@angular/flex-layout": "^13.0.0-beta.38",
-        "@angular/forms": "~13.4.0",
-        "@angular/material": "^13.3.9",
-        "@angular/platform-browser": "~13.4.0",
-        "@angular/platform-browser-dynamic": "~13.4.0",
-        "@angular/router": "~13.4.0",
-        "ngx-mat-timepicker": "^13.2.7",
-        "rxjs": "~6.6.0",
-        "tslib": "^2.6.0",
-        "zone.js": "~0.13.1"
-    },
-    "devDependencies": {
-        "@angular-devkit/build-angular": "^13.3.11",
-        "@angular/cli": "~13.3.7",
-        "@angular/compiler-cli": "~13.3.11",
-        "@types/jasmine": "~3.6.0",
-        "@types/node": "^12.11.1",
-        "dotty": ">=0.1.2",
-        "engine.io": ">=6.2.1",
-        "follow-redirects": ">=1.14.8",
-        "jasmine-core": "~3.8.0",
-        "json-schema": ">=0.4.0",
-        "karma": "~6.3.14",
-        "karma-chrome-launcher": "~3.1.0",
-        "karma-coverage": "~2.0.3",
-        "karma-jasmine": "~4.0.0",
-        "karma-jasmine-html-reporter": "^1.5.0",
-        "mpath": ">=0.8.4",
-        "typescript": "~4.6.4",
-        "node-forge": ">=1.3.0"
-    }
+  "name": "mist-webhook-monitor",
+  "version": "0.0.0",
+  "scripts": {
+    "ng": "ng",
+    "start": "ng serve --proxy-config ./proxy.conf.json --host 0.0.0.0",
+    "build": "ng build",
+    "watch": "ng build --watch --configuration development",
+    "test": "ng test"
+  },
+  "private": true,
+  "dependencies": {
+    "@angular/animations": "^14.3.0",
+    "@angular/common": "^14.3.0",
+    "@angular/compiler": "^14.3.0",
+    "@angular/core": "^14.3.0",
+    "@angular/forms": "^14.3.0",
+    "@angular/material": "^14.2.7",
+    "@angular/platform-browser": "^14.3.0",
+    "@angular/platform-browser-dynamic": "^14.3.0",
+    "@angular/router": "^14.3.0",
+    "rxjs": "~6.6.0",
+    "tslib": "^2.6.0"
+  },
+  "devDependencies": {
+    "@angular-devkit/build-angular": "^14.2.13",
+    "@angular/cli": "^14.2.13",
+    "@angular/compiler-cli": "^14.3.0",
+    "@types/jasmine": "~3.6.0",
+    "@types/node": "^12.11.1",
+    "dotty": ">=0.1.2",
+    "engine.io": ">=6.2.1",
+    "follow-redirects": ">=1.14.8",
+    "jasmine-core": "~3.8.0",
+    "json-schema": ">=0.4.0",
+    "karma": "~6.3.14",
+    "karma-chrome-launcher": "~3.1.0",
+    "karma-coverage": "~2.0.3",
+    "karma-jasmine": "~4.0.0",
+    "karma-jasmine-html-reporter": "^1.5.0",
+    "mpath": ">=0.8.4",
+    "typescript": "~4.6.4",
+    "node-forge": ">=1.3.0"
+  }
 }
